Add Validation of the new SugarCane module versus the old Fortran Sugar module (Apsim 7.7 official release)
Had to modify both simulations slightly due to features not being present in both Sugar and SugarCane modules.
Subsurface Irrigations not present in Sugar, and Irrigation with Solutes in it not present in SugarCane

Have to change the Input in the Datastore to the output file of the simulation in the validation dataset that you which to compare.

19b simulation still has some issues but I will come back and fix it later.