Add Validation of SugarCane module against actual observed data from experiments.
Added back in features such as subsurface irrigations that I could not do in the Fortran Validation.
Irrigation with Solutes is still missing. I will need to add this feature into the Irrigation module for ApsimX then come back and add it to the appropriate simulations in this validation set.