Get rid manager created reporting variables used to create graphs in old fortran validation simulations. These are just cruft in ApsimX.
Get rid of Harvest_Observed.txt because it is just a filtered version of All_Observed.txt (just the harvest observations) but I noticed that it was created incorrectly and the harvest dates are not the actual harvest dates.
If I ever need this again I will recreate it from All_Observed.txt , which will ensure it should be error free. It was only used for graphing purposes anyway.