Change Sugarcane Validation Harvesting Dates by one day (See below for reason) Added a copy of the original validation .apsimx file with the original correct harvest dates to make restoring faster.
I had to increase the harvest dates in every simulation by one day.
This is because the model does management at the start of everyday (ie. the crop is harvested at the start of every day)
But the "daily" reporting is the very last thing that happens each day.
Therefore if we set our harvest date to be the same as that in the observed data, then our predicted data at harvest will all be zeroed (because after the model harvests it zeros the values).
This is not a problem when we only report at "harvesting" instead of "daily" because the reporting occurs before the values are zeroed.
However we have to report daily because we have observation data during the crop growth not just at harvesting.