Fixed graph export to clipboard
Previously, graphs exported to the clipboard appeared all black. This
was due to the fact that OxyPlot's PngExporter was not using a
SolidPattern when used with certain versions of Mono.Cairo. To work
around this, we use its static Export method instead, which allows us
to force it to use a SolidPattern.