Decrement version number in some validation sets
Several validation sets have somehow ended up with a version number
which is higher than the current latest version. These need to be
reduced to the latest version now that it is illegal to open an .apsimx
file which has a version number greater than the latest version.

These files were