Fixed ExcelInput + JobRunnerMultiProcess
There were two problems preventing ExcelInput (and other IRunnables
which use a DataStore) from running with the multi process job runner.
First, if they attempt to do an Apsim.Find(this, typeof(IDataStore)),
they will end up using a DataStore, when they should be using a
StorageViaSockets. This is easily fixed by changing them to use a link,
and having the APSIMRunner resolve links for all non-RunSimulation
jobs.

The second problem is that the JobRunnerMultiProcess assumes that all
jobs are RunSimulations, and it attempts to use their DataStore
property to write all data received via the socket. This fails if the
job is not a RunSimulation. The solution here was to use a fallback
Apsim.Find() method to attempt to find the DataStore if the job is an
IModel. Note that this will still fail if the job is not an IModel or a
RunSimulation. Still not ideal, but better.