ListView should fallback to default cell settings
In a gtk treeview's SetCellDataFunc, changing a cell's property (e.g.
strikethrough) will change that property for all other cells as well.
To workaround this, we need to explicitly set each property every time
we go through the method, to ensure that the correct settings are
applied to each cell.