Added decumulate function
The decumulate function is an IFunction similar to the accumulate
function, except that instead of adding a small delta each day, it
subtracts the delta from the accumulated value.

This is needed to mimic behaviour of a stress effect in the sorghum
model in old apsim. I couldn't find a simple way to do this in the PMF
so this function should do the trick.