Fixed potential race condition in job runner code
It's not feasible for all simulations to have a child DataStore added
at simulation runtime - what would its Parent be? Currently each
simulation changes the parent property to point to itself but when
simulations are run in parallel, the DataStore's parent could
potentially point to a different simulation, which allows
models/events to propagate across simulations.

The only reason the Simulation needs a child DataStore is for link
resolution. The solution is to pass a reference to the DataStore into
the Links object via its Services property. This will probably cause
problems if we ever try to add any [Link] fields to the DataStore.
There's also still the problem of one simulation being able to
potentially access data from another simulation while it is being run;
the simulation would have a reference to the DataStore which has a
reference to the top-level Simulations node via its Parent property.