Added dltTT to sorghum arbitrator
Sorghum calculates daily TT as TT * stress. However, stress can change
between phenology calculations (which occur at start of day) and
reporting (which occurs at end of day). The Stress factor in the TT
calculation can change during the day which results in the value of TT
shown in the report being different to the value used in phenology
calculations earlier in the day. This rather hacky workaround gets the
sorghum arbitrator to store the thermalTime value early in the day in
the DoPhenology event, before the Stress value has a chance to change.