Fixed bug in clem/html views
This bug occurs when we click on a node which uses more than one html
view. The html view needs to wait for the browser widget to load its
contents before applying theme-specific overrides (back/foreground
colours for example).

Waiting for the browser to load its contents is achieved by waiting for
the gtk event queue to be clear. However, when we begin this wait, the
cursor changed event which triggered the new view to be loaded has
still not been fully processed. This means that this event will be
recursively called while waiting for the events queue to be cleared,
which causes a number of problems.

This solution is to detach the cursor changed event from its handler
while the handler is running.