Implemented edit file from shell feature
This allows the user to specify an .apsimx file and a config file when
running Models.exe. The .apsimx file will not be run but instead, the
changes listed in the config file will be applied to the .apsimx file,
which will then be written to disk under the same filename.

The config file should contain lines of the form 'path = value'

e.g.

[Clock].StartDate = 2019-1-20
.Simulations.Sim1.Name = SimulationVariant35
.Simulations.Sim2.Enabled = false
.Simulations.Sim1.Paddock.Soil.Thickness[1] = 50

Notes:

- Command line arguments should look like:

Models.exe file.apsimx /Edit /path/to/config/file.conf

- Relative paths will be resolved to the first match. ie
[Clock].StartDate will match the first clock found in the file.

- Dates can be specified as yyyy-mm-dd or mm/dd/yyyy.

- Strings should not be quoted

- Array indices will be interpted as 1-indexed (😡). So the first
element in the array should have index 1 in the config file.

- The file will be upgraded to the latest file version as part of this
process.