Fixed bug in R component
So it seems that due to a bug in R(?) that when we load a package via
library('asdf', lib.loc = '/packages/'), lib.loc is ignored in certain
circumstances (maybe while loading dependencies or if dependencies are
installed to a different location?). The simplest solution is to add
this path to .libPaths for the current session.