Fixed manager parameter parsing for IPlant properties
propertyType.IsAssignableFrom(typeof(IModel)) won't work here because
propertyType can be and often is an interface like IPlant which is
not assignable from IModel but its implementations are.

The result of this commit is that manager floating point parameters
will be incorrectly parsed if they omit the digits before the decimal
place - e.g. ".1". However, Apsim.Get(x, ".1") will no longer return
the root simulations node.