Fixed problem with IE browser not requesting enough space
Ok so for reasons I don't understand, the main widget's size request
seems to be in some cases smaller than the browser's size request. As a
result, the HTMLView will sometimes overlap with other widgets because
the HTMLView's size request is actually smaller than the space used by
the browser. In this scenario I would have expected the browser widget
to be cut off at the limits of the main widget's gdk window, but this
doesn't happen - perhaps due to a limitation or oversight in the gtk
socket component which we use to wrap the browser widget.

Either way, this little hack seems to correct the problem.