Improved SoilCrop profile grid if the soil crop is a factor
A soil crop profile grid doesn't display correctly if its parent is not
a physical node. This can happen if the soil crop is a factor or under
replacements. If under replacements, all bets are off. Otherwise, we
find an ancestor which is a simulation generator (experiment,
simulation, morris, etc.) and search for a physical node somewhere
under the simulation generator.