Fixed Sorghum SDRatio function
The SDRatio in sorghum is a moving average over the last five days. We
want to start accumulating values immediately after sowing but it looks
like there's a bug in the MovingAverageFunction which prevents that -
it never receives a phase changed event for sowing so it never starts
accumulating values, which causes SDRatio (and thus dltSlaiWater) to
always be zero which we don't want.

The fix is to just bite the bullet and start accumulating values at
germination (the first event received by the function).