Fixed sorghum/c4maize leafNoEffective calculation
Sorghum and maize calculate this differently so the calculation kind of
has to be in the simulations tree (in IFunctions), but we need to use
leafNoAtAppearance in the calculation which is specific to each culm.

Therefore I've implemented this stupid hack wherein we specify
leaf number at appearance as a public property of the structure object,
immediately before calling the calculation which uses this property.