Added json utility functions to rename models
I've added a new utility function to rename a model, which should
help prevent occurrences of sibling models with the same name from
occuring via the converter, as this causes a multitude of problems.

Manual name assignment should be avoided in the future, in favour of
using these methods. Would be nice if there was a way to enforce this..