Fixed typo in agpasture observed file (RadCliff -> Radcliffe)
If you look in the agpasture datastore, you'll notice that the Annual
table is missing, despite there being a PredictedObserved model called
Annual. The problem is that the simulation names in apsim are called
"RadcliffeLocationxxxx" but in the observed data they're called
"RadcliffLocationxxxx". I've changed the simulation names in the .xlsx
file.