Fixed typo in potato observed data (PlantAndHarvest -> PlantandHarvest)
The potato FinalYieldData predicted/observed table was matching the
data on the "PlantandHarvest.Script.FinalTag" variable name. There was
a typo in the case in the observed data which allowed the two tables to
be joined successfully and the predicted/observed table was created, but
the table didn't include the PlantandHarvest.Script.FinalTag variable
because they didn't match exactly.

This was causing problems for the performance tests.