Fixed intellisense not appearing in property presenter
OK so sometimes the HTMLView's web browser will steal focus. There is a
hack in the HTMLView code which manually gives focus back to the
toplevel window in this situation, but apparently creating the
intellisense popup is enough to trigger this hack if there is a
HTMLView onscreen. This is a problem because giving the focus back to
the main window will take focus away from the intellisense popup which
causes the intellisense popup to disappear. The workaround is to wait
for all Gtk events to process, and then recreate the intellisense popup
if need be.