Added a hack to allow for zone names containing "from" and "to"
Previously, if a zone name contained "from" or "to" then the zone name
would not be automatically reported because the report component would
treat the word "to" as if it were part of the aggregation syntax. This
change will cause report to ignore the "from" or "to" keywords if only
one is provided.

Note this doesn't help at all if the model name contains "as" or "of".