Fixed insertion of intellisense items in series filter
The series filter completion options will typically contain the word
which was used to generate completion options. E.g. the user might type
"Maize." and a suggestion might be "Maize.AboveGround.N". Previously,
we would end up with "MaizeMaize.AboveGround.N." which is obviously not
ideal.

To correct the problem, we now remove the trigger word ("Maize"
in this case) from the front of the completion suggestion and then
remove leading periods from the completion option if the text before
the cursor ends with a period.