Improved insertion of completion items in EditorView
Previously, inserting an intellisense suggestion could cause the editor
to scroll back to the top. This seems to be caused by setting the Text
property of the text editor widget. The workaround I've implemented
here is to record the scroll position before editing the text property
and then reset the scroll position manually after modifying this
property.