Profile graphs should not be added into simulations tree
Adding this graph into the simulations tree has some unfortunate side
effects such as making the graph visible in the UI in certain
situations (e.g. after enabling/disabling the node). We don't *need* to
add it to the tree for the graph to work (setting the soil as the
parent of the graph is sufficient).