Introduced an excel input model which reads from multiple sources
This differs from the existing ExcelInput component in that it will
read data from multiple .xlsx files and merge any sheets which exist in
all files. It also will not throw if any of the files don't exist.

This is needed because some of the observed data in the soybean
validation set is not public domain and will need to be encrypted in
version control. The idea is that the CI server and model developers
will decrypt the data when working on soybean but for other people the
validation set should still run seamlessly - the missing .xlsx file
will be ignored.