Fixed NFlow mechanism of finding source/destination solutes
This mechanism previously failed if another node of the wrong type but
correct name existed. This occurred when running a simulation with SWIM
which had a child called "Urea". A better solution would be to create a
new method in the Apsim API which allows us to find a model of a given
name *and* type, but that's probably beyond the scope of this issue.