Fixes to SimpleGrazing
Several fixes to the model.
The operations component doesn't need to come after the SimpleGrazing in the tree anymore (the order doesn't matter anymore).
Fixed the bug created when SimpleGrazing generates a "manager". This was confusing one of the graphs name with a parameter.
Deleted the option of TargetMassLength. Now lengths thresholds for rotations can be set for both TargetMass and FlexibleGrazing.