Fixed sorghum/maize accumTT variable
In old apsim, accumulated TT is calculated using the start-of-day
thermal time delta, which can actually be different to the end-of-day
delta if we progress to a new phase on that day, because the stress
can differ from phase to phase. In new apsim, however, the accumulated
TT is calculated from thermalTime.Value(), which is effectively an end-
of-day delta and as such can sometimes be different to the value used
in old apsim.