Fixed C4Maize grain stress function
In sorghum this stress factor only accumulates after flag leaf, but in
maize it's supposed to start accumulating at floral init and stop at
flowering. Previously, C4Maize was using the same start/stop as sorghum
which was causing differences in outputs vs old apsim. There will be a
follow-up commit to this, in which I update the .out files.