Force manager scripts to recompile after a binary clone
For reasons I don't fully understand, manager components with similar
code can sometimes end up with the same script object after binary
de/serialization. I'm not sure if this is a bug in the runtime, or a
result of our abuse of the runtime by creating multiple different
classes of the same name in the same namespace.

Either way, this patch corrects the issue by forcing manager scripts
to recompile after being cloned.