Should be able to copy a simulations node into the GUI
When this happens, the previous intended behaviour was to add the first
child of the simulations node. This wasn't working (unsure if it ever
did). I've fixed this by moving the test for simulations node into the
Structure class. Otherwise the UI will be out of sync with the actual
state of the file/model.