Fixed table merging in ExcelMultiInput
This component is designed to read observed data from multiple input
files and merge the tables if the files contain sheets with the same
name. The algorithm is to read a file, write the tables to the
datastore, and then move onto the next file, merge with any existing
tables in the datastore, write the table, and move onto the next file.

The problem is that observed data often contains simulation names,
whereas the tables already written to the datastore contain
SimulationID. When we merge the tables, we need to remember to drop
the SimulationID column, because SimulationID will be null for the data
which has just been read in from the .xlsx file.