Fixed c4maize dmPlantMax TT accumulation.
Previously, we used phenology.AccumTT, but the problem with this
variable is that it accumulates the end-of-day TT, whereas the old
maize phenology model uses start-of-day TT. These two values are
usually the same, but in the case of maize are different on the day
in which we enter the flowering phase (the stress is different).

This change adds a new variable (AccumTT) to the sorghum arbitrator
which mimics the accumulated pre-phenology dltTT.

The overall effect of this bug is pretty small, so this change could
(should?) be reverted at some point.