PropertyPresenter should refresh the grid after model is changed
This is needed specifically for the PredictedObserved model, although
it will likely be useful for others as well. The "field name to match
on" property's list of allowed values is dependent on the value of the
names of the predicted and observed tables. If these changed, the list
of valid field names to match on will also (probably) change.