Implemented automatic empty row deletion in profile grids
Any empty rows in the view for physical/soilwat/initialN/etc will now
be deleted whenever the user clicks away to a different node in the
tree, the rationale being that users probably don't want rows
disappearing as they're trying to type/copy/paste data here and there.

Without this, users inevitably leave empty rows in their soil profiles
which leads to a proliferation of NaNs in the soil object.