When copying properties from one model to another, ignore properties with XmlIgnore
Many properties (particularly in the new WaterBalance model) will throw
if links have not been resolved. This is a problem because after
deserializing a resource model, we need to copy all property values
from the 'official' resource, and links have not been resolved at this
point in time.

This solution is to ignore all properties with a jsonignore or
xmlignore attribute - these properties would not have been serialized
as part of the resource model anyway.