Fixed timing of TT target calculation in sorghum culms code
In old apsim, this variable is updated at the start of the day, before
phenology calculations occur, but after the call to calcFinalLeafNo();
Need to make sure that this doesn't happen until after calculaion
leafNo in new apsim as well.

Note that the only time that this matters is on the day we transition
to the endJuvenile phase, and only for cultivars who have a photoperiod
effect.