Fixed graphs with >1 checkpoint but no vary by settings
If an .apsimx file contains multiple checkpoints set to appear on
graphs, but a graph doesn't have either colour, line type or marker
type set to vary by something (e.g. vary by simulation name) then only
the data for the current checkpoint will appear on the graph. This is a
common situation when graphing data for a single simulation.