Fixed some discrepancies in C4Maize test set
Lines starting with asterisk are response from Jason after checking REMS.
Still need to sort out the African simulations.

-------------------------------------------------------------------------------
Angelo Gatton 99

- The initial N for the medium (N70) and high (N300) N treatment simulations
  were originally different.

* REMS has the NO3 in the soil as identical for all 3 treatments.
-------------------------------------------------------------------------------
UQ Gatton

All simulations are the same (with different cultivars) except for Barker,
which is different to the other sims in a few ways:

- End date is 20 March, whereas every other UQGatton sim ends on 16 March
- Initial NO3 in soil is way higher than the other sims
- More fertiliser is applied

I've changed barker to be identical to the other simulations (except for
cultivar, obviously) in new apsim and old apsim.

* Barker should be the same
-------------------------------------------------------------------------------
DRK

- All simulations apply 300kg fertiliser on sowing, except for:
  - Simulations sowing in August use 250kg
  - Hycorn 40 in August sim uses 200kg
- All sims which sow in March (except for the gh_5009 sim) have different
  sowing depth (70mm) and auto irrigation settings (FASWDepth = 1800,
  FASWThreshold = 0.75). Every other sim sows at 30mm and uses FASWDepth = 600,
  FASWThreshold = 0.9.

I've removed these discrepancies from both old and new apsim - that is, all
simulations now apply 300kg of fertiliser and use the same initial N, sowing
depth, and auto irrigation settings.

* REMS has them all the same. It doesn't specify the amount of N used at sowing - it should be unlimited I think.

-------------------------------------------------------------------------------
K841

- Maize_K841_K841QK6D applies 40mm irrigation on 1983-11-26, whereas every
  other dry simulation applies 15mm on this date.
- Maize_K841_K841HY9W has auto irrigation enabled but no other wet simulation
  has this
- K841 experiment in the released apsimx validation set sows Dekalb_XL82 in
  every simulation...why?

* REMS has the irrigation for Maize_K841_K841QK6D as 40mm on that date - different to the others.
* Maize_K841_K841HY9W should not be set to auto irrigate
* Dekalb_XL82 was planted using population of 3,5,7,9
* Other cultivars were only planted at populatio of 7
* Maize is only using the Dekalb_XL82 data
-------------------------------------------------------------------------------
K842

- Maize_K842_K842HY9W has auto irrigation enabled but no other wet (or dry for
  that matter) simulation has this
* Maize_K842_K842HY9W should not be auto.
-------------------------------------------------------------------------------
BMW1

No real discrepancies noted vs old apsim, however this experiment is quite
different to BMW1 in the official maize validation set in next gen. The sowing
rules in particular are oddly different (in row spacing and occasionally in
plant density)

-------------------------------------------------------------------------------
JMW2Katumani

The plant density in sowing rules for the high N treatments was different to
the plant density in the equivalent low-N treatment. I've changed these to be
identical to that of the equivalent low-N treatment.

- j2ka06: no change
- j2ka07: 2.2 -> 2.17
- j2ka08: 4.24 -> 4.19
- j2ka09: 6.13 -> 6.26
- j2ka10: 7.28 -> 7.33

-------------------------------------------------------------------------------
JMW2Kiboko

The plant density in sowing rules for the high N treatments was different to
the plant density in the equivalent low-N treatment. I've changed these to be
identical to that of the equivalent low-N treatment.

- j2ki06: no change
- j2ki07: 2.17 -> 2.18
- j2ki08: 4.39 -> 4.23
- j2ki09: 6.54 -> 6.22
- j2ki10: 7.63 -> 7.91

The initial N in the soils is different in the high-N treatments. I'm assuming
this is deliberate.

j2ki05 originally had curve runoff number set to 80. Every other j2k* sim has
this number set to 65. Is this an oversight? For now I have changed it to 80 in
both old and new for consistency's sake.

-------------------------------------------------------------------------------