Fixed flicker in HTMLView in dark mode on windows
When the browser loads and creates its body for the first time,
its BackColor and ForeColor properties are reset to their default
values (white/black). This causes a flicker when in dark mode.
To work around this, we embed some css into the markup when we
first load the document. This is a pretty gnarly workaround so
may need to be tweaked if there's some strange html passed into here.