Manager scripts should *never* recompile after the simulation starts
Otherwise we can get into a situation where the previously-compiled
script is still hooked up to the clock's events but is otherwise
inacessible, and the new script is acessible but not wired up to
receive any events.

The reason we need to force rebuilding the scripts is to work around a
weird issue (might be a bug in the runtime due to our abuse of it) in
which the script objects change during de/serialization. Forcing the
manager components to rebuild their scripts after this occurs is my
current workaround. This occurs once per simulation.

I've moved this force rebuilding into SimulationDescription, where it
will occur after the binary clone, but before the events are connected.