Remove [NonSerialized] attribute from Manager.compiledCode
If this is NonSerialized, then its value won't be copied when the
simulation is sent to a multi process job, which means that the manager
will recompile its script. However, this occurs *after* factors are
applied to the simulation, which means that if a factor modifies a
property in a manager script, the chagned value won't be retained if
the property has a [JsonIgnore] or [XmlIgnore] attribute. It's
debatable whether we want to actually suport this (ab)use of manager
scripts, but it was previously working so until it starts causing
problems, I'm happy to maintain this behaviour.