Edit file feature should work on resource models
Previously, changes made using the /Edit feature on Models.exe would
not persist if the affected model was a resource model (or a descendant
of one). This is because the full structure of the model wasn't being
serialized. I have changed this so that the full resource model will be
serialized after applying changes via /Edit. Only resource models whose
parameters have been changed will be affected.