GUI should remain responsive while running post-simulation tools
This change affects the gui when the user right clicks on the datastore
and clicks "refresh". The overall behaviour is similar to when running
simulations - there is a progress bar, status message and stop button.
Unfortunately, the progress bar is always at 100% because the
post-simulation tools aren't actually IRunnables and aren't included in
the progress calculation, but this is still an improvement to what we
had previously, where the gui would just freeze until they finish.