Fixed a bug in ModelReplacement which caused models to be incorrectly recreated
Previously, whenever we replaced a model via Replacements (and possibly
by factors too), we would call model.OnCreated() for all descendants of
the parent of the model being created. This means that if you replace a
cultivar and a report, and the cultivar is applied first, then all
descendants of the report's parent (which is usually the zone) would be
recreated. This is unnecessary - we really only want the newly added
model and its descendants to have their OnCreated() method called.