Slightly improve detection of expressions in report code
This is not really related to the previous change but it's pretty small so
I thought I'd throw it in here. Basically, any usage of ^ as a power operator
in an expression in the report was previously causing a fatal error, because
this line here didn't consider the caret to be a valid operator (which it is).