Factor and CompositeFactor should implement IReferenceExternalFiles
These models often reference external files. The method I've implemented for detecting
file references is to look at all the factor values and check if any are file paths.

This is required for CroptimizR, where we need to copy any referenced files to the
temp directory alongside the generated .apsimx file.