Change type of Configuration.Font from Pango.FontDescription to string
In gtk3 builds, we will load the theme by setting the GTK_THEME
environment variable *before* calling Application.Init(). However, this
means we cannot reference any GTypes.

Setting the theme variable correctly requires us to check Configuration
instance to see if the dark theme is enabled. Since this happens before*
Application.Init(), this means that Configuration cannot referene any
GTypes - such as Pango.FontDescription. This solution is to change the
font variable to a string.

Warning: this may damage the config file if you run a version of apsim
which includes this commit, and then switch to an older build from
before this commit. It won't cause a fatal error, but user settings
such as font, theme, recent files, etc may be reset to their default
values.