Fixed a bug which occurred in gtk3 builds involving grids
The bug occurred when selecting a different model in the simulations
tree while a cell in a GridView was in edit mode. Basically, the column
was being disposed and then a stray call to TreeView.GetCursor() was
returning a column whose cells had already been disposed which led to
some interesting results.