Fixed resource model's parameters not being correctly overriden
If a resource model (e.g. maize) is copied into replacements, and its
property values changed, these changed values will be overriden with the
'accepted' values from the official maize model when the simulation is
run, because the model's resource name is not null. This can be manually
rectified by editing the json, but such an intervention shouldn't be
necessary.

This solution is to manually set resource name to null in
ModelReplacement when replacing the model. Hopefully this doesn't have
any adverse effects...