Attempt to compile disabled managers when opening a file
If we don't compile these managers when opening the file, they're not
compiled until we click on the manager in the GUI and then click away
again.

This solution is to attempt to compile all managers when opening the
GUI. If a disabled manager fails to compile, then we suppress the error
messages.