Fixed adding resoursce model to replacements again
When creating a resource model, only read in the resource if a resource
name is provided AND either

1. It has no children OR
2. It's not under replacements

If the model is not under replacements, we always want to read in the
resource model. If it is under replacements, only read in the resource
model if it has no children. When we first instantiate a resource model
under replacements we will want to populate it with the version stored
as a resource. At all other times - when it has children - we will want
to *not* read the resource model and overwrite the potentially modified
children. What a mess this is.