Added a simpler property view which doesn't use a grid
This new property presenter/view serves the same basic purpose as the
old PropertyPresenter + GridView combo, but instead of laying out
properties in a grid, they are laid out in a series of label + input
rows, where the input is determined by the DisplayAttribute on the
property, and can be a text input, dropdown, checkbox, etc.

Still not finished, or used by any models, but as a proof of concept it
seems to be working ok.