Changed windows installer to uncheck 'create desktop icon' by default
Note that if the installer detects a previously-installed version of
APSIM, then it will check which settings were used for the previous
installation, and use the same values. So if in a previous installation
the user enabled the 'create desktop icon' option, then that option will
be enabled by default in this installation.

This change only affects the default value the *first* time APSIM is
installed.