Improve right clicking on multiple rows in list view
When the user right clicks on a row on a TreeView, if multiple rows were
previously selected, they get unselected as part of the mouse down
event, but we only make the popup menu appear in the mouse up event.
This means that they right click on several selected rows and have the
selection disappear, which is not really the expected behaviour.