Added converter to change zoom/location of all maps to default
The old zoom will no longer make sense to the new mapping widget.
Retaining the old coordinates for the center of the map can lead to
unexpected results when changing the zoom by itself, so I've changed the
center of the map as well. This is a one-time change which will reset
all maps back to the default zoom/center locations. (Most of them don't
have sensible values anyway).