Allow for .csv weather files
When using a .csv file for weather data, a second file must be
specified, which contains constants such as latitude, tav, amp, etc
(Latitude is the only one which is mandatory). The constants file must
be formatted as zero or more lines of the form:

name = value

Units may be optionally specified in parentheses after the value.
Anything following an exclamation mark (!) is treated as a comment. So
it should be possible to copy all of the constants out of a regular .met
file and paste them into a new file and it should "just work".

Note that there's currently no support for specifying units in the .csv
file. I've also added an example (Examples/CsvWeather.apsimx) and a new
page to the docs. I've also moved the "using excel for weather data"
page on the site into a new subdirectory called met, which also contains
the docs for the new .csv weather feature.