Fixed summary view 'jump to simulation log' w. markdown view
This is a really crude hack. The problem is that the textview widget
needs to be a direct child of a scrolledwindow, otherwise the
scroll_to_x functions in the textview api don't work. Unfortunately,
changing the markdown view to add each widget under its own
scrolledwindow will probably have unintended consequences for other
views.

The solution I've implemented here is to have the summary view
(which 'owns' the markdown view) modify the widget structure of the
markdown view, iff the markdown view contains only a single textview.
This assumption may break if someone writes a model which outputs some
fancy markdown to the summary file (e.g. an image). Another potential
solution might be to change summaryview to inerhit from markdown view,
although I'm not sure that this really makes sense in an OO mindset.