Added a hack to improve sizing of tables in autodocs
This is a really nasty cludge to work around the fact that
AutoDocumentation.ParseTextForTags doesn't pick up on markdown/html and
insert the appropriate tags - it always just puts the text into a
paragraph tag. So if, for example, you have a markdown table, this
function will just create a paragraph tag and store the raw markdown as
a paragraph, rather than creating a table tag. The problem with this
approach is that we already look through all table tags later on, in
order to fix the default sizing. My solution was to look through all
paragraph tags and check if they contain tables, and if so, adjust the
sizing - yuck. This autodocs code really needs some tidying up - it's
really difficult to follow. This would be a big job though.