Property UI should always show selected value in dropdowns
Sometimes, the Property UI is unable to fetch a list of valid options
for a dropdown - e.g. when the property is a table name, but the
simulation hasn't been run yet so the table doesn't exist yet.

In these scenarios, if the property already has a value, we want to show
that value as an option (the only option) in the dropdown.