Removed some EventSubscribe attributes from scripts in pinus and euc
These file contained some manager scripts which contained this:

[EventSubscribe("DoManagement")]
public void DoThin(double x) { ... }

This method signature is not valid for a subscriber to the DoManagement
event. The reason this worked previously was because the event
subscription code ignored public methods, and so it didn't attempt to
hook this function up to the DoManagement event. I've changed this code
though so that public methods will also be considered as potential event
subscriber candidates, which caused pinus and euc test sets to fail due
to the presence of these invalid subscribers.

I fixed it by removing the EventSubscribe attribute, which will cause it
to do exactly what it was doing previously. Unsure how much of this was
intentional.