Manager UI should be more tolerant of errors in script
Previously, errors in the manager script could prevent the properties
tab from being populated after fixing said errors. This patch should
rectify the issue by allowing SimplePropertyPresenter to work with a
null model (it will just hook up to the view, allowing
RefreshModel(IModel) to be later called to connect it to a non-null
model after the errors in the script have potentially been resolved.