Removed OutputPath option from CroptimizR
The new behaviour is to always copy the output data (generated into a
temp directory) into the same directory as the .apsimx file. The
individual file names will all be prefixed with the name of the
CroptimizR node, so that multiple CroptimizRs running in parallel (ie
siblings in the same .apsimx file) don't interfere with each other.