Fixed some oddities in agpasture factorial setups (check this)
Cromwell had an unused physical child, which had no corresponding line
in the composite factor. This particular composite factor overrides the
soil in the simulation, and the soil in the factor contains an
identical physical node anyway, so I just deleted the unused physical
node.

SouthKaipara had a disabled soil which I've removed.

WinchmoreIrrig had an unused operations node which was probably
intended to override the operations node of the same name in the base
simulation. To maintain the previous behaviour, I've removed this
unused operations node.