Fixed tracker strings in bar graphs
The tracker string is the textbox which appears when holding down the
left mouse button while hovering over a graph series. It shows the name
of the series and the x/y values at the mouse location. By default it
will interpolate if the mouse is between x-values, but in a bar graph
scenario, when we're using a category axis, it really doesn't make a lot
of sense to get an x-value via interpolation, or to show the upper/lower
bounds of the bar.

To improve this display, I've changed the meaning of the
TrackerFormatString in the ColumnXYSeries class, so that it doesn't show
anything to do with upper/lower bounds, and to use labels for display if
any are available (e.g. to show Jan instead of 1).

To anyone seeking to change this further: make sure you test the graphs
in the weather UI (e.g. radiation graph), and the objective function
graph in the croptimizr example.